+++ gt_manage/miniprogram/pages/PageMan/upGoods/upGoods.js
@@ -1,18 +1,72 @@
 // miniprogram/pages/PageMan/upGoods/upGoods.js
+import { CloudFn } from '../../../utils/CloudFn'
+const cloudFn = new CloudFn()
 Page({
 
   /**
    * 页面的初始数据
    */
   data: {
-
+    upGoodsArray: [], // 上架商品列表
   },
 
   /**
    * 生命周期函数--监听页面加载
    */
   onLoad: function (options) {
+    this._loadData()
+  },
 
+
+  // 获取全部上架商品
+  _loadData: function() {
+    wx.showLoading({
+      title: '加载中...',
+      icon:'none'
+    })
+    cloudFn.$callFn({
+      data: {
+        name: 'CloudAPIBase',
+        body: {
+          fn: 'get',
+          base: 'shop-goods',
+          is_where: false,
+          where_data: {shelf:true}
+        }
+      }
+    }).then(res => {
+      wx.hideLoading()
+      this.data.upGoodsArray = JSON.parse(res.resp_data).data
+      this.setData({
+        upGoodsArray: this.data.upGoodsArray
+      })
+    })
+  },
+
+
+  // 收到下架通知， 下架
+  onShelf: function(event) {
+    const { id } = event.detail
+    wx.showLoading({
+      title: '下架中...',
+      icon:'none'
+    })
+    cloudFn.$callFn({
+      data: {
+        name: 'CloudAPIBase',
+        body: {
+          fn: 'update',
+          base: 'shop-goods',
+          where_data: {_id:id},
+          update_data: {
+            shelf: false
+          }
+        }
+      }
+    }).then(res => {
+      wx.hideLoading()
+      this._loadData()
+    })
   },
 
   /**