+++ gt_manage/miniprogram/pages/Tabbar/manage/manage.js
@@ -1,4 +1,6 @@
 // miniprogram/pages/Tabbar/manmage/manage.js
+import { CloudFn } from '../../../utils/CloudFn'
+const cloudFn = new CloudFn()
 Page({
 
   /**
@@ -53,7 +55,7 @@ Page({
    * 生命周期函数--监听页面加载
    */
   onLoad: function (options) {
-
+   
   },
 
   // 去对应页